Create Linux, Determine If You Are Using Root User
This will determine if you are running a script as root user. 

It checks your User ID to see if it is zero. Sine root is zero if the number is not equal to zero, it will kick you out. 

This is script should be at the beginning of all other scripts to ensure they run properly.